Able to display really basic model data in GUI.
Tester: in order to initially display what was in the model, all
operations done on the model had to be done before sending it to the
view. Now we should be able to see what is stored in the model upon
startup.

Part: Overrode the toString() method to print out all relevant part
info.

PartsInventoryModel: added a getInventory() method which returns a
List<Part> object.

PartsInventoryView: added an updatePanel() method which destroys the
current inventory list and pulls the latest data from the model to be
put on display. Had to use a DefaultListModel object to work with the
JList object (could have used a String[] array, but now we don't have to
deal with indices when adding elements).

PartsInventoryController: when Add is clicked, view.updatePanel() is
called, as well as view.repaint();