Part: refined exception data to print user friendly error messages.
PartView: added an "error message" label which gives the user feedback
in the GUI based on exceptions thrown or other errors encountered when
trying to manipulate individual parts.

PartsInventoryController: added a ListSelectionEvent handler which
reports all of the currently selected parts in the inventoryView. This
is used in conjunction with the Delete function: every selected part can
be deleted (one or more at a time).

PartsInventoryModel: refined sort methods so that subsequent calls would
toggle between ascending or descending sorts on the respective field.

PartsInventoryPanel: not sure if I made any significant changes here...
it didn't seem to be used in the View.

PartsInventoryView: added buttons and panels, registered some additional
listeners, and started making size and bounds associations with related
elements in the GUI layout objects.